### 2017-09-24 #415 @whedon
[PRE REVIEW]: micompm - A MATLAB/Octave toolbox for multivariate independent comparison of observations
**Submitting author:** @fakenmc (<a href="http://orcid.org/0000-0002-8487-5837">Nuno Fachada</a>)
**Repository:** <a href="https://github.com/fakenmc/micompm" target ="_blank">https://github.com/fakenmc/micompm</a>
**Version:** v1.0.0
