### 2017-09-28 #415 @cMadan
@fakenmc, your submission appears to be in good order--detailed examples, paper.md of appropriate form, and a substantive toolbox. Do you have any suggestions for reviewers?
