### 2017-09-29 #415 @fakenmc
Hi @cMadan , looking at JOSS's list of reviewers, I compiled the following list of potential reviewers for our submission:

@Kevin-Mattheus-Moerman
@fgnievinski
@effigies
@arokem
@pboesu
@ahwillia
@jordigh

Thanks in advance,

Nuno
