### 2018-01-30 #430 @jordigh
Okay, I worked through the tutorial. I think I can tick off all the checkboxes except for the following:

* I don't see a clear indication of how to seek support or report issues. I assume you expect to use github issues, but I don't see this stated anywhere.
* I don't see who the intended audience is. There is more statistics here than I am personally comfortable with. I was running a bit blind here. It looks like a fairly sophisticated statistical audience is intended.

I worked with Octave. I noticed some slightly different outputs for some of the tutorial examples, including some NaNs instead of some very small numbers. I don't know where the rounding errors came from.

I was unable to generate the LaTeX output as shown there. For some reason, none of the tickmarks by tikz would show up, although I could see the axes. It would help to have a complete source code of the intended LaTeX file to compile.
