### 2018-01-30 #430 @fakenmc
Thanks, I'll address these issues shortly.
