### 2018-01-31 #430 @fakenmc
I've worked through the raised points as follows:

* Added support and issue reporting information to the Contributing guidelines.
* Clarified who's the intended audience in README.md, User Guide and the submitted paper.
* Added a template LaTeX document for quick testing table generation; also added a reference to this template in the tutorial.

Concerning the small differences between MATLAB and Octave results. As described in the User Guide, Octave uses a different way (when compared to MATLAB) of determining the statistics on some of the performed statistical tests. The outputs in the tutorial are from MATLAB, thus the stated differences. I've performed a large number of tests to determine if these MATLAB vs Octave differences are meaningful, but the p-values were consistently similar. I have also validated these conclusions with R. Thus one will get to the same conclusion (on sample similarity or otherwise) using either software.

I believe to have addressed the raised points. If there is anything else I can do please let me know.
