### 2018-03-28 #430 @fakenmc
@cMadan , the DOI for the latest version of micompm is 10.5281/zenodo.931838 .

Thanks in advance
